Change entitlementId type from int to Long

+++ src/main/java/iss/nus/edu/sg/leave_application_processing_system/model/LeaveEntitlement.java
@@ -11,7 +11,7 @@ import jakarta.persistence.Table;
 public class LeaveEntitlement {
 
 	@Id
-	private int entitlementId;
+	private Long entitlementId;
 	private Employee employeeId;
 	@Enumerated (EnumType.STRING)
 	private LeaveType leavetype;
